Web Scraping Project with Price Comparison
This project is a project that compares the prices of a product on many sites and offers the most affordable prices. With web scraping, data of laptop products were obtained from 4 websites. Web scraping is the process of automatically extracting data from a website. This is done using a program or software to retrieve data from a particular web page. Web scraping is also referred to as scraping or pulling data from sources such as HTML or API from the website. This data is photos, name, brand, model, product price and product features. The pulled products and their data are added to the mangodb database and displayed in a web view. Django was used for the web view.